Testing Updates
-Mocha and Chai updated
-Got initial instance of testing setup working 
-Two tests built in behaviors
