+++ package.json
@@ -8,7 +8,7 @@
         "@microsoft/microsoft-graph-types": "^1.32.0",
         "@pnp/buildsystem": "^2.3.1",
         "@types/adal-angular": "^1.0.1",
-        "@types/chai": "^4.2.15",
+        "@types/chai": "^4.2.21",
         "@types/chai-as-promised": "^7.1.3",
         "@types/core-js": "^2.5.4",
         "@types/es6-promise": "^3.3.0",
@@ -17,7 +17,7 @@
         "@types/gulp-replace": "^0.0.31",
         "@types/jsonwebtoken": "^8.5.0",
         "@types/lodash.clonedeep": "^4.5.6",
-        "@types/mocha": "^8.2.1",
+        "@types/mocha": "^9.0.0",
         "@types/node": "^14.14.34",
         "@types/pump": "^1.1.1",
         "@types/sharepoint": "^2016.1.9",
@@ -37,7 +37,7 @@
         "https-proxy-agent": "^5.0.0",
         "jsonwebtoken": "^8.5.1",
         "lodash.clonedeep": "^4.5.0",
-        "mocha": "^8.3.2",
+        "mocha": "^9.1.0",
         "msal": "^1.4.6",
         "node-fetch": "^2.6.1",
         "prettyjson": "^1.2.1",
