+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
     "compilerOptions": {
-        "target": "ES5",
+        "target": "ES2018",
         "module": "commonjs",
         "composite": true,
         "declaration": true,
